src/test/java/be/intecbrussel/blogapplication/controllers/HomeControllerTest.java
package be.intecbrussel.blogapplication.controllers;

import be.intecbrussel.blogapplication.model.User;
import be.intecbrussel.blogapplication.services.UserService;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.setup.MockMvcBuilders;
import org.springframework.ui.Model;

import java.security.Principal;


import static org.junit.jupiter.api.Assertions.*;
import static org.mockito.ArgumentMatchers.anyLong;
import static org.mockito.Mockito.when;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.view;

class HomeControllerTest {

    @Mock
    UserService userService;

    @Mock
    Model model;

    @Mock
    Principal principal;

    MockMvc mockMvc;

    HomeController controller;

    @BeforeEach
    void setUp() {
        MockitoAnnotations.openMocks(this);
        controller = new HomeController(userService);
        mockMvc = MockMvcBuilders.standaloneSetup(controller).build();
    }

    @Test
    void guestUserRoot() throws Exception {

        mockMvc.perform(get("/"))
                .andExpect(status().isOk())
                .andExpect(view().name("home"));
    }

    @Test
    void loggedInUserRoot() {

        User user = new User();

        when(userService.findById(anyLong())).thenReturn(user);

        String viewName = controller.root(principal, model);

        assertEquals("index",viewName);
    }

    @Test
    void login() throws Exception {

    }

    @Test
    void userIndex() {
    }

    @Test
    void handleError() {
    }

    @Test
    void getErrorPath() {
    }
}